backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
boto3==1.34.34
botocore==1.34.34
python-dotenv==1.0.0
gunicorn==21.2.0
